requirements.txt
python-telegram-bot==20.3
requests==2.31.0
httpx
python-dotenv